fix: Remove excessive CF_NOESCAPE and CS_NOESCAPE
Analysed all occurrences of CF_NOESCAPE and NS_NOESCAPE and removed
where the annotation seemed to be excessive, for example when the
annotated block has been enqueued in a queue to execute in another loop.

Such a block, may have a an optimised placement as
`NSConcreteGlobalBlock` on stack on a callee side. This happen on
non-Apple clang Apple at least in version 7 (see [1]).

This is slightly different than fully disabling the blocks (see [2],
[3], and [4]), while trying to preserve as much of the NOESCAPES where
it seems to server a purpose.

This partially reverts commit 697acf9b0054093c0d7db6e381310daf0cc9da38.

* src/mac.c (mac_foreach_window_1, mac_foreach_window, mac_map_keymap):
  Restore CF_NOESCAPE.
* src/macappkit.m (mac_within_gui_and_here, mac_autorelease_loop):
  Restore CF_NOESCAPE.
* src/macterm.h (mac_foreach_window, mac_map_keymap) : Restore
  CF_NOESCAPE.
* src/macappkit.h (enumerateChildWindowsUsingBlock): Harmonise
  NS_NOESCAPE placement.
* src/macappkit.m (enumerateChildWindowsUsingBlock): Harmonise
  CF_NOESCAPE placement.

[1] https://github.com/NixOS/nixpkgs/issues/127902#issuecomment-1698365843
[2] https://github.com/jdtsmith/emacs-mac/issues/43
[3] https://github.com/jdtsmith/emacs-mac/pull/48
[4] https://github.com/jdtsmith/emacs-mac/issues/69